server/requirements.txt
blis==0.7.7
catalogue==2.0.7
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
cymem==2.0.6
en-core-web-sm==3.2.0
Flask==2.1.1
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
langcodes==3.3.0
libretranslatepy==2.1.1
lxml==4.8.0
MarkupSafe==2.1.1
murmurhash==1.0.6
numpy==1.22.3
packaging==21.3
pathy==0.6.1
preshed==3.0.6
pydantic==1.8.2
pyparsing==3.0.8
pyspellchecker==0.6.3
requests==2.27.1
smart-open==5.2.1
spacy==3.2.4
spacy-legacy==3.0.9
spacy-loggers==1.0.2
srsly==2.4.2
thinc==8.0.15
tqdm==4.64.0
translate==3.6.1
typer==0.4.1
typing-extensions==4.1.1
urllib3==1.26.9
wasabi==0.9.1
Werkzeug==2.1.1
zipp==3.8.0
